requirements.txt
aiofiles==0.7.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
elasticsearch==7.15.2
fastapi==0.68.1
frozenlist==1.2.0
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-resources==5.2.2
iniconfig==1.1.1
itsdangerous==2.0.1
Mako==1.1.5
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.5
python-multipart==0.0.5
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.1
urllib3==1.26.6
uvicorn==0.15.0
yarl==1.7.2
zipp==3.5.0
